app/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
#requests
# numpy==1.24.3
# openai==1.3.0
# azure-identity
# azure-core
# huggingface-hub
# transformers
# langchain
# torch
# together
# dotenv